Updated Pins For Hardware Design
Leaving PWM Pins open for another day.

Board orientation will be right side (out pin) is at the SD card, and the Top is where all the analog pins are.

Also removed potentiometer from design to allow students to input this value in the file.

Updated SD card protocol to:

    /* Format:  REQUEST_RATE_SEC;SECRET_SSID;SECRET_PASS;SERVER_NAME;IO_USERNAME;IO_GROUP;IO_FEED_KEY;IO_KEY */
    /* Example: 2;Issam's iPhone;Pass123;io.adafruit.com;AOMUser1;AOMGroupA;AOMGroupAFeed1;AoMIoKeYsEcReT123 */